# unused-param,unused-receiver: refactor configure func