The PR refactors the `configure` function in both `unused-param` and `unused-receiver`:

- change to use `blankIdentifierRegex` that is compiled once;
- reduce nested `if`s;
- add tests and benchmarks.

## Benchmarks

Benchmarks show that version in this PR has slightly fewer allocations.

```sh
$ git switch master
$ go test -benchmem -run=^$ -bench "^BenchmarkUnused.*" github.com/mgechev/revive/test -count=10 > old.txt
$ git switch refactor/unused-configure
$ go test -benchmem -run=^$ -bench "^BenchmarkUnused.*" github.com/mgechev/revive/test -count=10 > new.txt
$ benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: github.com/mgechev/revive/test
cpu: Apple M1 Pro
                 │   old.txt   │              new.txt               │
                 │   sec/op    │   sec/op     vs base               │
UnusedParam-8      671.0µ ± 1%   673.0µ ± 3%       ~ (p=0.436 n=10)
UnusedReceiver-8   211.9µ ± 8%   207.7µ ± 4%       ~ (p=0.063 n=10)
geomean            377.0µ        373.9µ       -0.83%

                 │   old.txt    │               new.txt               │
                 │     B/op     │     B/op      vs base               │
UnusedParam-8      325.5Ki ± 0%   323.7Ki ± 0%  -0.58% (p=0.000 n=10)
UnusedReceiver-8   80.22Ki ± 0%   78.09Ki ± 0%  -2.65% (p=0.000 n=10)
geomean            161.6Ki        159.0Ki       -1.62%

                 │   old.txt   │              new.txt               │
                 │  allocs/op  │  allocs/op   vs base               │
UnusedParam-8      8.322k ± 0%   8.290k ± 0%  -0.38% (p=0.000 n=10)
UnusedReceiver-8   1.896k ± 0%   1.864k ± 0%  -1.69% (p=0.000 n=10)
geomean            3.972k        3.931k       -1.04%
```